change qual score sum to uint16_t

+++ Genrich.h
@@ -209,7 +209,7 @@ typedef struct read {
   uint8_t alnLen;   // length of alignment array
   Aln* alnR2;       // array of alignments for R2 (discordant alns only)
   uint8_t alnLenR2; // length of alnR2 array (discordant alns only)
-  int qual;         // sum of quality scores
+  uint16_t qual;    // sum of quality scores
   bool first;       // true -> R1; false -> R2 (singleton alns only)
   float score;      // min. alignment score
   float scoreR2;    // min. alignment score for R2 (discordant alns only)
